[REFACTOR] tcp: Remove separation between Sender and Receiver
