+++ src/inetstack/protocols/layer4/tcp/established/mod.rs
@@ -10,9 +10,7 @@ mod congestion_control_state;
 pub mod ctrlblk;
 mod delivery_state;
 mod flow_control_state;
-mod receiver;
 mod rto;
-mod sender;
 
 //======================================================================================================================
 // Imports
@@ -32,8 +30,6 @@ use crate::{
                     ctrlblk::{ConnectionManagementState, ControlBlock, State},
                     delivery_state::DeliveryState,
                     flow_control_state::FlowControlState,
-                    receiver::Receiver,
-                    sender::Sender,
                 },
                 header::TcpHeader,
                 SeqNumber,
@@ -143,15 +139,14 @@ impl SharedEstablishedSocket {
             _ => (),
         };
 
-        let sender = Sender::new(sender_seq_no);
-        let receiver = Receiver::new(
+        let delivery = DeliveryState::new(
+            sender_seq_no,
             receiver_seq_no,
             receiver_seq_no,
             ack_delay_timeout_secs,
             receiver_window_size_bytes,
             receiver_window_scale_bits,
         );
-        let delivery = DeliveryState::new(sender, receiver);
 
         let flow_control = FlowControlState::new(
             sender_seq_no,
@@ -194,7 +189,7 @@ impl SharedEstablishedSocket {
 
         let now = self.runtime.now();
         let mut layer3_endpoint = self.layer3_endpoint.clone();
-        Receiver::receive(&mut self.control_block, &mut layer3_endpoint, tcp_hdr, buf, now);
+        DeliveryState::receive(&mut self.control_block, &mut layer3_endpoint, tcp_hdr, buf, now);
     }
 
     // This coroutine runs the close protocol.
@@ -218,10 +213,10 @@ impl SharedEstablishedSocket {
         // 2. Wait for FIN and FIN ack.
         let mut me2 = self.clone();
         let mut me3 = self.clone();
-        let wait_for_fin = pin!(me3.control_block.delivery.receiver.wait_for_fin().fuse());
+        let wait_for_fin = pin!(me3.control_block.delivery.wait_for_fin().fuse());
         let mut runtime = self.runtime.clone();
         let mut layer3_endpoint = self.layer3_endpoint.clone();
-        let push_fin_and_wait_for_ack = pin!(Sender::push(
+        let push_fin_and_wait_for_ack = pin!(DeliveryState::push(
             &mut me2.control_block,
             &mut layer3_endpoint,
             &mut runtime,
@@ -255,7 +250,7 @@ impl SharedEstablishedSocket {
         // 1. Send FIN and wait for ack before closing.
         let mut runtime = self.runtime.clone();
         let mut layer3_endpoint = self.layer3_endpoint.clone();
-        Sender::push(
+        DeliveryState::push(
             &mut self.control_block,
             &mut layer3_endpoint,
             &mut runtime,
@@ -270,11 +265,11 @@ impl SharedEstablishedSocket {
     pub async fn push(&mut self, bufs: ArrayVec<DemiBuffer, MAX_BATCH_SIZE_NUM_PACKETS>) -> Result<(), Fail> {
         let mut runtime = self.runtime.clone();
         let mut layer3_endpoint = self.layer3_endpoint.clone();
-        Sender::push(&mut self.control_block, &mut layer3_endpoint, &mut runtime, bufs).await
+        DeliveryState::push(&mut self.control_block, &mut layer3_endpoint, &mut runtime, bufs).await
     }
 
     pub async fn pop(&mut self, size: Option<usize>) -> Result<ArrayVec<DemiBuffer, MAX_BATCH_SIZE_NUM_PACKETS>, Fail> {
-        self.control_block.delivery.receiver.pop(size).await
+        self.control_block.delivery.pop(size).await
     }
 
     pub fn endpoints(&self) -> (SocketAddrV4, SocketAddrV4) {
@@ -288,7 +283,7 @@ impl SharedEstablishedSocket {
         let mut me = self.clone();
         let acknowledger = async_timer!("tcp::established::background::acknowledger", async {
             let mut layer3_endpoint = me.layer3_endpoint.clone();
-            Receiver::acknowledger(&mut me.control_block, &mut layer3_endpoint).await
+            DeliveryState::acknowledger(&mut me.control_block, &mut layer3_endpoint).await
         })
         .fuse();
         pin_mut!(acknowledger);
@@ -297,7 +292,7 @@ impl SharedEstablishedSocket {
         let retransmitter = async_timer!("tcp::established::background::retransmitter", async {
             let mut layer3_endpoint = me2.layer3_endpoint.clone();
             let mut runtime = me2.runtime.clone();
-            Sender::background_retransmitter(&mut me2.control_block, &mut layer3_endpoint, &mut runtime).await
+            DeliveryState::background_retransmitter(&mut me2.control_block, &mut layer3_endpoint, &mut runtime).await
         })
         .fuse();
         pin_mut!(retransmitter);
@@ -306,7 +301,7 @@ impl SharedEstablishedSocket {
         let sender = async_timer!("tcp::established::background::sender", async {
             let mut layer3_endpoint = me3.layer3_endpoint.clone();
             let mut runtime = me3.runtime.clone();
-            Sender::background_sender(&mut me3.control_block, &mut layer3_endpoint, &mut runtime).await
+            DeliveryState::background_sender(&mut me3.control_block, &mut layer3_endpoint, &mut runtime).await
         })
         .fuse();
         pin_mut!(sender);
